milestone3_modifications_and_updates
