+++ apps/tasks/models.py
@@ -16,8 +16,3 @@ class Comment(models.Model):
     owner = models.ForeignKey(User, on_delete=models.CASCADE, related_name='comments')
 
 
-class TimeLog(models.Model):
-    start_timer = models.DateTimeField(null=True)
-    end_timer = models.DateTimeField(null=True)
-    task = models.ForeignKey(Task, on_delete=models.CASCADE, db_index=True)
-    owner = models.ForeignKey(User, on_delete=models.CASCADE, db_index=True)
