+++ apps/tasks/views.py
@@ -7,10 +7,9 @@ from rest_framework.generics import GenericAPIView, get_object_or_404
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 
-from .models import Task, Comment, TimeLog
+from .models import Task, Comment
 from .serializers import TaskCreateSerializer, TaskSerializer, \
-    TaskUpdateSerializer, TaskUpdateStatusSerializer, CommentsSerializer, TaskSearchSerializer, TaskItemSerializer, \
-    TimelogSerializer
+    TaskUpdateSerializer, TaskUpdateStatusSerializer, CommentsSerializer, TaskSearchSerializer, TaskItemSerializer
 from ..users.models import User
 from ..users.serializers import GetUsersSerializer
 
@@ -121,19 +120,3 @@ class TaskSearchView(generics.ListAPIView):
     search_fields = ['title']
 
 
-class TimeLogViewSet(ModelViewSet):
-    queryset = TimeLog.objects.all()
-    serializer_class = TimelogSerializer
-
-    def get_serializer_class(self):
-        if self.action == 'create':
-            return TaskCreateSerializer
-        return super().get_serializer_class()
-
-    def perform_create(self, serializer):
-        serializer.save(owner=self.request.user)
-
-    # @action(methods=['GET'], detail=False, serializer_class=TimelogSerializer)
-    # def timelogs(self, request):
-    #     timelogs = TimeLog.objects.filter(owner=self.request.user)
-    #     return Response(TimelogSerializer(timelogs, many=True).data)
